Initial commit for core-edge
catchup
-------
 Store copy and transaction pulling, similar to the way slaves catch up with
 master in the traditional HA setup.

discovery
---------
Uses Hazelcast to discover an initial cluster topology both for Core and Edge
instances. This discovery layer is used to prompt RAFT cluster changes in Core as
instances leave and join, and used as a topology manager for Edge instances.

raft
----
An implementation of a distributed log kept consistent by RAFT. While there
are several packages in here, the key ones are:
  * log - where the log implementations for the RAFT protocol are kept
  * replication - machinery for replicating entries in the raft log
  * membership - where RAFT group membership is maintained (see above for
  initial setup)

server
------
 The server package is where all the code is brought together. It also provides
 some shared infrastructure for sending messages, dealing with settings etc.